feat(gateway): implement configuration and serial communication for MQ7 sensor
