+++ gateway/src/constants.rs
@@ -1,0 +1,7 @@
+pub const DEFAULT_TARGET: &str = "127.0.0.1:9000";
+pub const DEFAULT_INTERVAL_MS: u64 = 5000;
+pub const DEFAULT_ACK_TIMEOUT_MS: u64 = 3000;
+pub const DEFAULT_EXPECTED_ACK: &str = "ack:success";
+pub const DEFAULT_PAYLOAD_SUCCESS: &str = "success";
+pub const DEFAULT_SERIAL_BAUD: u32 = 115200;
+pub const DEFAULT_SERIAL_TIMEOUT_MS: u64 = 1200;
